Completed error validations on not found
