+++ app/javascript/components/app_error.vue
@@ -1,0 +1,13 @@
+<template lang="pug">
+  #error
+    h2 Sorry but this author is gone
+    hr
+    router-link(to='/') back to authors list
+</template>
+
+<script>
+export default {  }
+</script>
+
+<style scoped>
+</style>
